// src/app/models/master/masDepartmentData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface MasDepartmentData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    /** รหัสคณะ/หน่วยงาน(ห้ามซ้ำ) */
    depCode?: string;
    /** PK auto run (dep_id)  */
    depId?: number;
    /** ตัวย่อคณะ/หน่วยงาน (ภาษาอังกฤษ) */
    depNameAbbrEn?: string;
    /** ตัวย่อคณะ/หน่วยงาน (ภาษาไทย) */
    depNameAbbrTh?: string;
    /** ชื่อเต็มคณะ/หน่วยงาน (ภาษาอังกฤษ) */
    depNameEn: string;
    /** ชื่อย่อคณะ/หน่วยงาน (ภาษาอังกฤษ) */
    depNameShortEn?: string;
    /** ชื่อย่อคณะ/หน่วยงาน (ภาษาไทย) */
    depNameShortTh?: string;
    /** ชื่อเต็มคณะ/หน่วยงาน (ภาษาไทย) */
    depNameTh: string;
    /** fk (dep_parent_id) */
    depParentId?: number;
    /** สถานะจัดการเรียนการสอน */
    educationStatus?: boolean;
    first?: number;
    mode?: string;
    rowNum?: number;
    size?: number;
}